+++ ksqlDb.RestApi.Client/KSql/Query/Context/Options/KSqlDbProviderOptions.cs
@@ -5,5 +5,7 @@ namespace ksqlDb.RestApi.Client.KSql.Query.Context.Options
   public abstract class KSqlDbProviderOptions
   {
     internal JsonSerializerOptions JsonSerializerOptions { get; set; } = KSqlDbJsonSerializerOptions.CreateInstance();
+
+    public bool DisposeHttpClient { get; set; } = true;
   }
 }